Warning not to delete books

+++ src/main/java/com/example/librarymanagernqc/ManagementInterface/Document/DocumentController.java
@@ -5,6 +5,7 @@ import com.example.librarymanagernqc.ManagementInterface.Document.AddBook.AddBoo
 import com.example.librarymanagernqc.ManagementInterface.Document.BookInformation.BookInformationController;
 import com.example.librarymanagernqc.database.Controller.BookDatabaseController;
 import com.example.librarymanagernqc.database.DAO.BookDAO;
+import com.example.librarymanagernqc.database.DAO.BorrowedListDAO;
 import com.jfoenix.controls.JFXButton;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -45,6 +46,7 @@ public class DocumentController {
     private TextField searchField;
 
     private BookDAO bookDAO = new BookDAO();
+    private BorrowedListDAO borrowedListDAO = new BorrowedListDAO();
 
     /**
      * all books list
@@ -184,12 +186,23 @@ public class DocumentController {
                         deleteButton.setCursor(Cursor.HAND);
                         deleteButton.setGraphic(deleteImage);
                         deleteButton.setOnAction(event -> {
-                            //lấy ô hiện tại đang chọn
+                            // Lấy sách hiện tại đang được chọn
                             Book book = getTableView().getItems().get(getIndex());
 
-                            // Xóa ô khỏi book list, database
-                            deleteBookFromList(book);
+                            // Kiểm tra nếu sách đang được mượn
+                            if (borrowedListDAO.isBookLoanExist(book.getId())) {
+                                // Hiển thị thông báo
+                                Alert alert = new Alert(Alert.AlertType.WARNING);
+                                alert.setTitle("Can't delete a book");
+                                alert.setHeaderText(null);
+                                alert.setContentText("The book cannot be deleted because it is being borrowed.");
+                                alert.showAndWait();
+                            } else {
+                                // Xóa sách khỏi danh sách và database
+                                deleteBookFromList(book);
+                            }
                         });
+
                     }
 
                     private final HBox buttonsBox = new HBox(10, editButton, deleteButton);
